+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Addressed failure to pickup changes to source and destination if not matched with dropdown list
 * Fixed bug where source or destination could be incorrectly reported in the busy frame
+* Made the router handle cargo changes more intelligently
 
 ## v1.8.1 2026-03-11